tools to render steps as function calls
